# [Optimisation] Improved performance by refactoring part of the code